pyproject.toml
[tool.poetry]
name = "sd-mqtt-printer-mac"
version = "0.1.0"
description = "Mac USB thermal printer MQTT client - Python port of ESP32 firmware"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src", from = "."}]

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.6.1"
python-escpos = "^3.0"
qrcode = "^7.4.2"
pillow = "^10.0.0"
pyusb = "^1.2.1"
python-dotenv = "^1.0.0"
psutil = "^5.9.0"
requests = "^2.31.0"
schedule = "^1.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"
pytest = "^7.4.0"

[tool.poetry.scripts]
start = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
